Login, initial functionality.
- Created LoginUI
- Created util class for inputHandling
- fixed typos in packages

Added
- LoginController
- LoginUI
- util classes
