--- Test/Controllers/AdminControllerTest.java
@@ -1,7 +1,0 @@
-package Controllers;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class AdminControllerTest {
-
-}
