--- Test/Controllers/UserControllerTest.java
@@ -1,12 +1,0 @@
-package Controllers;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class UserControllerTest {
-
-    //TODO
-    /// AdminControllerTests
-
-
-
-}
