--- Test/Repositories/CurrencyRepositoryTest.java
@@ -1,7 +1,0 @@
-package Repositories;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class CurrencyRepositoryTest {
-
-}
