+++ Test/controllers/AdminControllerTest.java
@@ -1,0 +1,5 @@
+package controllers;
+
+class AdminControllerTest {
+
+}
