+++ Test/controllers/UserControllerTest.java
@@ -1,0 +1,10 @@
+package controllers;
+
+class UserControllerTest {
+
+    //TODO
+    /// AdminControllerTests
+
+
+
+}
