+++ Test/repositories/CurrencyRepositoryTest.java
@@ -1,0 +1,5 @@
+package repositories;
+
+class CurrencyRepositoryTest {
+
+}
