--- src/Controllers/AdminController.java
@@ -1,16 +1,0 @@
-package Controllers;
-
-public class AdminController {
-
-
-    //TODO
-    /// add attributes, Constructor getters and setters
-
-    /// start method()
-    /// implement navigation functionality between service class and UI
-    /// switch call methods from services.
-
-
-    ///
-
-}
