--- src/Controllers/UserController.java
@@ -1,13 +1,0 @@
-package Controllers;
-
-public class UserController {
-
-    //TODO
-    /// add attributes, Constructor getters and setters
-
-    /// start method()
-    /// implement navigation functionality between service class and UI
-    /// switch call methods from services.
-
-
-}
