--- src/Interfaces/Controller.java
@@ -1,19 +1,0 @@
-package Interfaces;
-
-public interface Controller {
-
-
-    //TODO
-    /// start()
-    // switch calling respective service methods
-
-
-
-
-    /// handleMenuInput()
-    // check input
-
-
-
-
-}
