--- src/Interfaces/UIInterface.java
@@ -1,8 +1,0 @@
-package Interfaces;
-
-public interface UIInterface {
-
-
-    //TODO
-    /// printMenu() method
-}
