+++ src/Main.java
@@ -1,5 +1,8 @@
+import controllers.LoginController;
+
 public class Main {
     public static void main(String[] args) {
-        System.out.println("Hello world");
+        LoginController loginController = new LoginController();
+        loginController.start();
     }
 }
