--- src/Repositories/CurrencyRepository.java
@@ -1,16 +1,0 @@
-package Repositories;
-
-public class CurrencyRepository {
-
-    //TODO
-
-    // list<Currency>
-
-    //readList()
-
-    //getCurrency
-
-
-
-
-}
