--- src/Repositories/StockMarketRepository.java
@@ -1,14 +1,0 @@
-package Repositories;
-
-public class StockMarketRepository {
-
-    //TODO
-
-    // list<StockMarket>
-
-    //readFile()
-
-    //getStockMarket()
-
-
-}
