--- src/Repositories/TransactionRepository.java
@@ -1,16 +1,0 @@
-package Repositories;
-
-public class TransactionRepository {
-
-
-    //TODO
-
-    // List<Transaction>
-
-
-    //readFile()
-
-    //getTransaction()
-
-
-}
