--- src/Repositories/UserRepository.java
@@ -1,17 +1,0 @@
-package Repositories;
-
-public class UserRepository {
-
-
-    //TODO
-
-    //List<User>
-
-    // readFile()
-
-    // writeToFile()
-
-    //getUser()
-
-
-}
