--- src/Services/CurrencyService.java
@@ -1,18 +1,0 @@
-package Services;
-
-public class CurrencyService {
-
-
-
-    //TODO
-
-
-    // list<Currency>
-
-    // getCurrencyRate()
-
-
-
-    // kommunikation til controller
-
-}
