--- src/Services/LoginService.java
@@ -1,20 +1,0 @@
-package Services;
-
-public class LoginService {
-
-
-    private boolean isAdmin;
-    private boolean isLoggedIn;
-
-
-    //TODO
-
-    // checkUserInput()
-
-    // isMemberFunctionality()
-
-    // isAdminFunctionality()
-
-
-
-}
