--- src/Services/StockMarketService.java
@@ -1,19 +1,0 @@
-package Services;
-
-public class StockMarketService {
-
-
-    //TODO
-
-    //List<StockMarked>
-
-
-    // getStockMarket()
-
-    // getTicker()
-
-    // getPrice()
-
-    // getCurrency()
-
-}
