--- src/Services/UserService.java
@@ -1,20 +1,0 @@
-package Services;
-
-public class UserService {
-
-
-    //TODO
-
-    //findUserData()
-
-    // writeNewUserData() -> Admin
-
-    //findUserByEmail()
-
-    // sortUsersByName() -> andim eller alle?
-
-    // sortUsersByCash()
-
-    // getUserID()
-
-}
