--- src/UI/AdminUI.java
@@ -1,4 +1,0 @@
-package UI;
-
-public class AdminUI {
-}
