--- src/UI/LoginUI.java
@@ -1,4 +1,0 @@
-package UI;
-
-public class LoginUI {
-}
