+++ src/controllers/AdminController.java
@@ -1,0 +1,16 @@
+package controllers;
+
+public class AdminController {
+
+
+    //TODO
+    /// add attributes, Constructor getters and setters
+
+    /// start method()
+    /// implement navigation functionality between service class and UI
+    /// switch call methods from services.
+
+
+    ///
+
+}
