+++ src/controllers/UserController.java
@@ -1,0 +1,13 @@
+package controllers;
+
+public class UserController {
+
+    //TODO
+    /// add attributes, Constructor getters and setters
+
+    /// start method()
+    /// implement navigation functionality between service class and UI
+    /// switch call methods from services.
+
+
+}
