+++ src/interfaces/Controller.java
@@ -1,0 +1,19 @@
+package interfaces;
+
+public interface Controller {
+
+
+    //TODO
+    /// start()
+    // switch calling respective service methods
+
+
+
+
+    /// handleMenuInput()
+    // check input
+
+
+
+
+}
