+++ src/repositories/CurrencyRepository.java
@@ -1,0 +1,16 @@
+package repositories;
+
+public class CurrencyRepository {
+
+    //TODO
+
+    // list<Currency>
+
+    //readList()
+
+    //getCurrency
+
+
+
+
+}
