+++ src/repositories/StockMarketRepository.java
@@ -1,0 +1,14 @@
+package repositories;
+
+public class StockMarketRepository {
+
+    //TODO
+
+    // list<StockMarket>
+
+    //readFile()
+
+    //getStockMarket()
+
+
+}
