+++ src/repositories/TransactionRepository.java
@@ -1,0 +1,16 @@
+package repositories;
+
+public class TransactionRepository {
+
+
+    //TODO
+
+    // List<Transaction>
+
+
+    //readFile()
+
+    //getTransaction()
+
+
+}
