+++ src/repositories/UserRepository.java
@@ -1,0 +1,17 @@
+package repositories;
+
+public class UserRepository {
+
+
+    //TODO
+
+    //List<User>
+
+    // readFile()
+
+    // writeToFile()
+
+    //getUser()
+
+
+}
