+++ src/services/CurrencyService.java
@@ -1,0 +1,18 @@
+package services;
+
+public class CurrencyService {
+
+
+
+    //TODO
+
+
+    // list<Currency>
+
+    // getCurrencyRate()
+
+
+
+    // kommunikation til controller
+
+}
