+++ src/services/StockMarketService.java
@@ -1,0 +1,19 @@
+package services;
+
+public class StockMarketService {
+
+
+    //TODO
+
+    //List<StockMarked>
+
+
+    // getStockMarket()
+
+    // getTicker()
+
+    // getPrice()
+
+    // getCurrency()
+
+}
