+++ src/services/UserService.java
@@ -1,0 +1,20 @@
+package services;
+
+public class UserService {
+
+
+    //TODO
+
+    //findUserData()
+
+    // writeNewUserData() -> Admin
+
+    //findUserByEmail()
+
+    // sortUsersByName() -> andim eller alle?
+
+    // sortUsersByCash()
+
+    // getUserID()
+
+}
