+++ src/ui/AdminUI.java
@@ -1,0 +1,4 @@
+package ui;
+
+public class AdminUI {
+}
