+++ src/ui/UserUI.java
@@ -1,0 +1,4 @@
+package ui;
+
+public class UserUI {
+}
